@@ -1,5 +1,5 @@
 import React from 'react';
-import { HashRouter as Router, Route, Switch } from 'react-router-dom';
+import { HashRouter as Router, Route, Routes } from 'react-router-dom';
 import ActionsGui from './views/ActionsGui';
 import BrowserGui from './views/BrowserGui';
 import NotFound from './views/NotFound';
@@ -7,18 +7,32 @@ import OnScreenGui from './views/OnScreenGui';
 import OnTouchGui from './views/OnTouchGui';
 import RemoteGui from './views/RemoteGui';
 
+import Overlay from './components/common/Overlay/Overlay';
+import Stack from './components/common/Stack/Stack';
+import About from './views/About/About';
+
+function OnTouchAbout() {
+  return <Overlay>
+    <Stack style={{ maxWidth: '500px' }}>
+        Close
+      <About />
+    </Stack>
+  </Overlay>;
+}
+
 const App = () => (
   <Router>
-    <Switch>
-      <Route path="/onscreen" component={OnScreenGui} />
-      <Route path="/remote" component={RemoteGui} />
-      <Route path="/ontouch" component={OnTouchGui} />
-      <Route path="/actions" component={ActionsGui} />
+    <Routes>
+      <Route path="/onscreen" element={<OnScreenGui />} />
+      <Route path="/remote" element={<RemoteGui />} />
+      <Route path="/ontouch" element={<OnTouchGui />} />
+      <Route path="/ontouch/about" element={<OnTouchAbout />} />
+      <Route path="/actions" element={<ActionsGui />} />
       {/* Here, more GUI variations can be added. */}
-      {/* <Route path="/tablet" component={TabletGui} /> */}
-      <Route path="/" component={BrowserGui} />
-      <Route component={NotFound} />
-    </Switch>
+      {/* <Route path="/tablet" element={TabletGui} /> */}
+      <Route path="/" element={<BrowserGui />} />
+      <Route element={<NotFound />} />
+    </Routes>
   </Router>
 );
 
