@@ -1,7 +1,6 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 // AppContainer is a necessary wrapper component for hot module reloading
-import { AppContainer } from 'react-hot-loader';
 import { Provider } from 'react-redux';
 import { compose, createStore } from 'redux';
 import middleware from './api/Middleware';
@@ -19,20 +18,10 @@ const store = createStore(
   )
 );
 
-const render = (Component) => {
-  ReactDOM.render(
-    <Provider store={store} >
-      <AppContainer>
-        <Component />
-      </AppContainer>
-    </Provider>,
-    document.getElementById('root'),
-  );
-};
-
-render(App);
-
-// Hot Module Replacement API
-if (module.hot) {
-  module.hot.accept('./App', () => render(App));
-}
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(
+  <Provider store={store} >
+    <App />
+  </Provider>
+);
