@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { withRouter } from 'react-router-dom';
+import { useLocation } from "react-router-dom";
 import { setShowAbout, startConnection } from '../api/Actions';
 import { formatVersion, isCompatible, RequiredOpenSpaceVersion, RequiredSocketApiVersion } from '../api/Version';
 import BottomBar from '../components/BottomBar/BottomBar';
@@ -16,80 +16,73 @@ import '../styles/base.scss';
 import About from './About/About';
 import styles from './OnScreenGui.scss';
 
-class OnScreenGui extends Component {
-  constructor(props) {
-    super(props);
-    this.checkedVersion = false;
-    this.showFlightController = props.showFlightController;
-  }
+function OnScreenGui({ showFlightController, connectionLost, startConnection, version, showAbout, hideAbout }) {
+  let hasCheckedVersion = false;
+  let location = useLocation();
 
-  componentDidMount() {
-    this.props.startConnection();
-  }
+  React.useEffect(() => {
+    startConnection();
+  }, []);
 
-  checkVersion() {
-    if (!this.checkedVersion && this.props.version.isInitialized) {
-      const versionData = this.props.version.data;
-      if (!isCompatible(versionData.openSpaceVersion, RequiredOpenSpaceVersion)) {
-        console.warn(
-          'Possible incompatibility: \nRequired OpenSpace version: ' +
-          formatVersion(RequiredOpenSpaceVersion) +
-          '. Currently controlling OpenSpace version ' +
-          formatVersion(versionData.openSpaceVersion) + '.'
-        );
-      }
-      if (!isCompatible(versionData.socketApiVersion, RequiredSocketApiVersion)) {
-        console.warn(
-          "Possible incompatibility: \nRequired Socket API version: " +
-          formatVersion(RequiredSocketApiVersion) +
-          ". Currently operating over API version " +
-          formatVersion(versionData.socketApiVersion) + '.'
-        );
-      }
-      this.checkedVersion = true;
+  // Check the version
+  if (!hasCheckedVersion && version.isInitialized) {
+    const versionData = version.data;
+    if (!isCompatible(versionData.openSpaceVersion, RequiredOpenSpaceVersion)) {
+      console.warn(
+        'Possible incompatibility: \nRequired OpenSpace version: ' +
+        formatVersion(RequiredOpenSpaceVersion) +
+        '. Currently controlling OpenSpace version ' +
+        formatVersion(versionData.openSpaceVersion) + '.'
+      );
     }
+    if (!isCompatible(versionData.socketApiVersion, RequiredSocketApiVersion)) {
+      console.warn(
+        "Possible incompatibility: \nRequired Socket API version: " +
+        formatVersion(RequiredSocketApiVersion) +
+        ". Currently operating over API version " +
+        formatVersion(versionData.socketApiVersion) + '.'
+      );
+    }
+    hasCheckedVersion = true;
   }
 
-  reloadGui() {
+  function reloadGui() {
     location.reload();
   }
 
-  render() {
-    this.checkVersion();
-    return (
-      <div className={styles.app}>
-        { this.props.showAbout && (
-          <Overlay>
-            <Stack style={{ maxWidth: '500px' }}>
-              <Button style={{ alignSelf: 'flex-end', color: 'white' }} onClick={this.props.hideAbout}>
-                Close
-              </Button>
-              <About />
-            </Stack>
-          </Overlay>
-        )}
-        { this.props.connectionLost && (
-          <Overlay>
-            <Error>
-              <h2>Houston, we've had a...</h2>
-              <p>...disconnection between the user interface and OpenSpace.</p>
-              <p>Trying to reconnect automatically, but you may want to...</p>
-              <Button className={Error.styles.errorButton} onClick={this.reloadGui}>Reload the user interface</Button>
-            </Error>
-          </Overlay>
-        )}
-        <section className={styles.Grid__Left}>
-          <Sidebar />
-        </section>
-        <section className={styles.Grid__Right}>
-          <NodePopOverContainer />
-          <NodeMetaContainer />
-          <BottomBar showFlightController={this.props.showFlightController}/>
-          <KeybindingPanel />
-        </section>
-      </div>
-    );
-  }
+  return (
+    <div className={styles.app}>
+      { showAbout && (
+        <Overlay>
+          <Stack style={{ maxWidth: '500px' }}>
+            <Button style={{ alignSelf: 'flex-end', color: 'white' }} onClick={hideAbout}>
+              Close
+            </Button>
+            <About />
+          </Stack>
+        </Overlay>
+      )}
+      { connectionLost && (
+        <Overlay>
+          <Error>
+            <h2>Houston, we've had a...</h2>
+            <p>...disconnection between the user interface and OpenSpace.</p>
+            <p>Trying to reconnect automatically, but you may want to...</p>
+            <Button className={Error.styles.errorButton} onClick={reloadGui}>Reload the user interface</Button>
+          </Error>
+        </Overlay>
+      )}
+      <section className={styles.Grid__Left}>
+        <Sidebar />
+      </section>
+      <section className={styles.Grid__Right}>
+        <NodePopOverContainer />
+        <NodeMetaContainer />
+        <BottomBar showFlightController={showFlightController}/>
+        <KeybindingPanel />
+      </section>
+    </div>
+  );
 }
 
 const mapStateToProps = state => ({
@@ -107,9 +100,9 @@ const mapDispatchToProps = dispatch => ({
   }
 });
 
-OnScreenGui = withRouter(connect(
+OnScreenGui = connect(
   mapStateToProps,
   mapDispatchToProps,
-)(OnScreenGui));
+)(OnScreenGui);
 
 export default OnScreenGui;
